directory structure to match other components
